services/heroes/requirements.txt
starlette
uvicorn
asyncpg
python-dotenv
prometheus-client
